# db: add D/II LRU caches in snapshots