--- a/silkworm/db/datastore/snapshots/common/key_hasher.cpp
+++ b/silkworm/db/datastore/snapshots/common/key_hasher.cpp
@@ -14,18 +14,18 @@
    limitations under the License.
 */
 
-#include "bloom_filter_key_hasher.hpp"
+#include "key_hasher.hpp"
 
 #include <array>
 
 #include "../common/encoding/murmur_hash3.hpp"
 
-namespace silkworm::snapshots::bloom_filter {
+namespace silkworm::snapshots {
 
-uint64_t BloomFilterKeyHasher::hash(ByteView key) const {
+uint64_t KeyHasher::hash(ByteView key) const {
     std::array<uint64_t, 2> hash = {0, 0};
     encoding::Murmur3{salt_}.hash_x64_128(key.data(), key.size(), hash.data());
     return hash[0];
 }
 
-}  // namespace silkworm::snapshots::bloom_filter
+}  // namespace silkworm::snapshots