--- a/silkworm/db/datastore/snapshots/inverted_index_seek_query.hpp
+++ b/silkworm/db/datastore/snapshots/inverted_index_seek_query.hpp
@@ -30,41 +30,73 @@ namespace silkworm::snapshots {
 
 template <EncoderConcept TKeyEncoder>
 struct InvertedIndexSeekSegmentQuery {
-    explicit InvertedIndexSeekSegmentQuery(InvertedIndex entity)
-        : query_{std::move(entity)} {}
+    explicit InvertedIndexSeekSegmentQuery(InvertedIndex entity, InvertedIndexCache* cache)
+        : query_{std::move(entity)}, cache_{cache} {}
 
     using Key = decltype(TKeyEncoder::value);
 
     std::optional<datastore::Timestamp> exec(Key key, datastore::Timestamp timestamp) {
-        auto list_opt = query_.exec(std::move(key));
+        TKeyEncoder key_encoder;
+        key_encoder.value = std::move(key);
+        ByteView key_data = key_encoder.encode_word();
+        return exec_raw(key_data, timestamp);
+    }
+
+    std::optional<datastore::Timestamp> exec_raw(ByteView key, datastore::Timestamp timestamp) {
+        std::optional<InvertedIndexTimestamps> ts_pair;
+        uint64_t key_hash_hi{0};
+
+        if (cache_) {
+            if (std::tie(ts_pair, key_hash_hi) = cache_->get(key); ts_pair) {
+                if (ts_pair->requested <= timestamp) {
+                    if (timestamp <= ts_pair->found || ts_pair->found == 0) {
+                        return ts_pair->found;
+                    }
+                }
+            }
+        }
+
+        auto list_opt = query_.exec_raw(key);
         if (!list_opt) {
             return std::nullopt;
         }
 
         elias_fano::EliasFanoList32& list = *list_opt;
-        auto result = list.seek(timestamp);
-        return result ? std::optional{result->second} : std::nullopt;
+        const auto seek_result = list.seek(timestamp);
+        if (!seek_result) {
+            return std::nullopt;
+        }
+        const auto [_, higher_or_equal_ts] = *seek_result;
+
+        if (cache_) {
+            cache_->put(key_hash_hi, InvertedIndexTimestamps{.requested = timestamp, .found = higher_or_equal_ts});
+        }
+
+        return higher_or_equal_ts;
     }
 
   private:
     InvertedIndexFindByKeySegmentQuery<TKeyEncoder> query_;
+    InvertedIndexCache* cache_;
 };
 
 template <EncoderConcept TKeyEncoder>
 struct InvertedIndexSeekQuery {
     explicit InvertedIndexSeekQuery(
         const SnapshotRepositoryROAccess& repository,
-        std::function<InvertedIndex(const SnapshotBundle&)> entity_provider)
+        std::function<InvertedIndex(const SnapshotBundle&)> entity_provider,
+        std::function<InvertedIndexCache*()> cache_provider)
         : repository_{repository},
-          entity_provider_{std::move(entity_provider)} {}
+          entity_provider_{std::move(entity_provider)},
+          cache_provider_{std::move(cache_provider)} {}
 
     using Key = decltype(TKeyEncoder::value);
 
     std::optional<datastore::Timestamp> exec(Key key, datastore::Timestamp timestamp) {
         datastore::TimestampRange ts_range{timestamp, repository_.max_timestamp_available() + 1};
-        for (auto& bundle_ptr : repository_.bundles_intersecting_range(ts_range, /* ascending = */ true)) {
+        for (auto& bundle_ptr : repository_.bundles_intersecting_range(ts_range, /*ascending=*/true)) {
             const SnapshotBundle& bundle = *bundle_ptr;
-            InvertedIndexSeekSegmentQuery<TKeyEncoder> query{entity_provider_(bundle)};
+            InvertedIndexSeekSegmentQuery<TKeyEncoder> query{entity_provider_(bundle), cache_provider_()};
             auto result = query.exec(key, timestamp);
             if (result) {
                 return result;
@@ -76,6 +108,7 @@ struct InvertedIndexSeekQuery {
   private:
     const SnapshotRepositoryROAccess& repository_;
     std::function<InvertedIndex(const SnapshotBundle&)> entity_provider_;
+    std::function<InvertedIndexCache*()> cache_provider_;
 };
 
 }  // namespace silkworm::snapshots