--- a/silkworm/db/datastore/snapshots/snapshot_repository_ro_access.hpp
+++ b/silkworm/db/datastore/snapshots/snapshot_repository_ro_access.hpp
@@ -28,6 +28,7 @@
 #include "../common/step.hpp"
 #include "../common/timestamp.hpp"
 #include "common/util/iterator/map_values_view.hpp"
+#include "domain_cache.hpp"
 #include "segment_and_accessor_index.hpp"
 
 namespace silkworm::snapshots {
@@ -59,6 +60,8 @@ struct SnapshotRepositoryROAccess {
     };
 
     virtual ~SnapshotRepositoryROAccess() = default;
+
+    virtual DomainCache* domain_cache(const datastore::EntityName& name) const = 0;
 
     virtual size_t bundles_count() const = 0;