### Functionality

- Introduce Domain LRU cache (one for each entity) used in `DomainGetLatestQuery `
- Introduce InvertedIndex LRU cache (one for each entity) used in `InvertedIndexSeekQuery`
- Introduce `exec_raw` in segment queries to avoid multiple key encoding
- Move `BloomFilterKeyHasher` and rename as `KeyHasher` for reuse in snapshot LRU caches

### Performance

MacBookPro M1-Max 64GB RAM

- master: f22d3473dafbab954d8387914960c5d1074bfc47
```
% ./run_perf_tests.py -m 2 -p pattern/mainnet/stress_test_eth_getLogs_15M.tar -r 5 -t 1000:20 -y eth_getLogs
Performance Test started
[1. 1] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=1.68s]
[1. 2] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=1.058s]
[1. 3] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=1.032s]
[1. 4] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=980.002ms]
[1. 5] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=1.053s]
Performance Test completed successfully.
```

- this branch: 540eb5382011efe189e73d536ddd86791b0b75b8
```
./run_perf_tests.py -m 2 -p pattern/mainnet/stress_test_eth_getLogs_15M.tar -r 5 -t 1000:20 -y eth_getLogs
Performance Test started
Test repetitions: 5 on sequence: 1000:20 for pattern: pattern/mainnet/stress_test_eth_getLogs_15M.tar
[1. 1] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=744.158ms]
[1. 2] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=588.391ms]
[1. 3] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=615.164ms]
[1. 4] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=597.833ms]
[1. 5] daemon: executes test qps: 1000 time: 20 -> [R=100.00% max=594.024ms]
Performance Test completed successfully.
```